requirements.txt
celery
redis
python-dotenv
